Version 20200505 update
Add --makeone option to create a single man file in the package root.
It will still need to be added to debian/<package>.manfiles.
This option might be useful if the auto-generated file needs tweaking
or fails to build, eg for lack of a Build-Depends.

Extend and reformat the --help output.

Remove local variable creation, which is not POSIX compliant.
Functions are run in subshells instead.

Add a GPL licence header.